refactor animated height to own component using HOC. Move and update tests
