+++ src/animated-wrapper.js
@@ -2,18 +2,14 @@ import React from "react";
 import PropTypes from "prop-types";
 import { Motion } from "react-motion";
 
-const AnimatedWrapper = ({ animateWrapper, children, height, onRest }) =>
-  animateWrapper ? (
-    <Motion defaultStyle={{ height: 0 }} style={{ height }} onRest={onRest}>
-      {children}
-    </Motion>
-  ) : (
-    children({ height: "auto" })
-  );
+const AnimatedWrapper = ({ children, height, onRest }) => (
+  <Motion defaultStyle={{ height: 0 }} style={{ height }} onRest={onRest}>
+    {children}
+  </Motion>
+);
 
 AnimatedWrapper.propTypes = {
   children: PropTypes.func,
-  animateWrapper: PropTypes.bool,
   height: PropTypes.oneOfType([PropTypes.number, PropTypes.object]),
   onRest: PropTypes.func
 };
